+++ ftests.py
@@ -1,5 +1,6 @@
 import unittest
-from functional_tests.FYLM141111001 import FunctionalTests
+from functional_tests.FYLM141111001 import FunctionalTests as FYLM141111Tests
+from functional_tests.single import FunctionalTests as SingleTests
 
 
 if __name__ == '__main__':