+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## [1.1.2] - 2015-10-09
+### ADDED
+- `Image` objects now have a `frame_number` attribute.
+
 ### CHANGED
 - `Image` objects now directly subclass Numpy arrays.
 - Refactored code to permit parsing of different versions of ND2s, which will allow us to add support for NIS Elements 3.x.