+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-<<<<<<< HEAD
 ## [1.1.2] - 2015-10-09
 ### ADDED
 - `Image` objects now have a `frame_number` attribute.
+- `Nd2` can be used as a context manager.
 
 ### CHANGED
 - `Image` objects now directly subclass Numpy arrays.
@@ -11,8 +11,6 @@
 - The `data` attribute is no longer needed since `Image` is now a Numpy array.
 - The `image_sets` iterator will be removed in the near future. You should implement this yourself.
 
-=======
->>>>>>> origin/master
 ## [1.1.1] - 2015-09-02
 ### FIXED
 - Images returned by indexing would sometimes be skipped when the file contained multiple channels